Make all ts/*.html files include only module.js and module.css
