+++ ts/card-info/card-info.html
@@ -3,9 +3,7 @@
     <head>
         <meta charset="utf-8" />
         <meta name="viewport" id="viewport" content="width=device-width" />
-        <link href="card-info-base.css" rel="stylesheet" />
         <link href="card-info.css" rel="stylesheet" />
-        <script src="../js/vendor/bootstrap.bundle.min.js"></script>
         <script src="card-info.js"></script>
     </head>
     <body>
