+++ ts/change-notetype/change-notetype.html
@@ -3,9 +3,7 @@
     <head>
         <meta charset="utf-8" />
         <meta name="viewport" id="viewport" content="width=device-width" />
-        <link href="change-notetype-base.css" rel="stylesheet" />
         <link href="change-notetype.css" rel="stylesheet" />
-        <script src="../js/vendor/bootstrap.bundle.min.js"></script>
         <script src="change-notetype.js"></script>
     </head>
     <body>
