+++ ts/congrats/congrats.html
@@ -3,7 +3,6 @@
     <head>
         <meta charset="utf-8" />
         <meta name="viewport" id="viewport" content="width=device-width" />
-        <link href="congrats-base.css" rel="stylesheet" />
         <link href="congrats.css" rel="stylesheet" />
         <script src="congrats.js"></script>
     </head>
