+++ ts/deck-options/deck-options.html
@@ -3,9 +3,7 @@
     <head>
         <meta charset="utf-8" />
         <meta name="viewport" id="viewport" content="width=device-width" />
-        <link href="deck-options-base.css" rel="stylesheet" />
         <link href="deck-options.css" rel="stylesheet" />
-        <script src="../js/vendor/bootstrap.bundle.min.js"></script>
         <script src="deck-options.js"></script>
     </head>
     <body>
