+++ ts/graphs/graphs.html
@@ -3,7 +3,6 @@
     <head>
         <meta charset="utf-8" />
         <meta name="viewport" id="viewport" content="width=device-width" />
-        <link href="graphs-base.css" rel="stylesheet" />
         <link href="graphs.css" rel="stylesheet" />
         <script src="graphs.js"></script>
     </head>
